requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
numpy==2.0.1
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pytesseract==0.3.10
requests==2.32.3
tesseract==0.1.3
urllib3==2.2.2
Wand==0.6.13
